update request processing for `PRE` and refactor long types to `uint64_t`.

+++ include/memory_request.h
@@ -10,9 +10,7 @@
 
 #include "common.h"
 
-void memory_request_init(MemoryRequest_t *memoryRequest, unsigned long time,
-                         uint8_t core, uint8_t operation,
-                         unsigned long long address);
+void memory_request_init(MemoryRequest_t *memoryRequest, uint64_t time, uint8_t core, uint8_t operation, uint64_t address);
 char *memory_request_to_string(MemoryRequest_t *memory_request);
 
 #endif
